refactor(Google Java Style Convention)
