+++ src/main/java/baseball/checker/Result.java
@@ -6,24 +6,24 @@ public class Result {
     private static final String STRIKE_MESSAGE = "스트라이크";
     private static final String BALL_MESSAGE = "볼";
     private static final String NOTHING_MESSAGE = "낫싱";
-    private int strike;
-    private int ball;
+    private int strikeCount;
+    private int ballCount;
 
     public Result() {
-        this.strike = 0;
-        this.ball = 0;
+        this.strikeCount = 0;
+        this.ballCount = 0;
     }
 
     public void strike() {
-        strike++;
+        strikeCount++;
     }
 
     public void ball() {
-        ball++;
+        ballCount++;
     }
 
     public boolean isAnswer() {
-        if (strike == BASEBALL_NUMBERS_SIZE) {
+        if (strikeCount == BASEBALL_NUMBERS_SIZE) {
             clear();
             return true;
         }
@@ -31,23 +31,23 @@ public class Result {
     }
 
     public void printResult() {
-        if (ball > 0) {
-            System.out.print(ball + BALL_MESSAGE);
+        if (ballCount > 0) {
+            System.out.print(ballCount + BALL_MESSAGE);
         }
-        if (ball > 0 && strike > 0) {
+        if (ballCount > 0 && strikeCount > 0) {
             System.out.print(' ');
         }
-        if (strike > 0) {
-            System.out.print(strike + STRIKE_MESSAGE);
+        if (strikeCount > 0) {
+            System.out.print(strikeCount + STRIKE_MESSAGE);
         }
-        if (ball == 0 && strike == 0) {
+        if (ballCount == 0 && strikeCount == 0) {
             System.out.print(NOTHING_MESSAGE);
         }
         System.out.println();
     }
 
     public void clear() {
-        strike = 0;
-        ball = 0;
+        strikeCount = 0;
+        ballCount = 0;
     }
 }
