+++ src/main/java/baseball/checker/types/EachNumberResultType.java
@@ -1,4 +1,4 @@
-package baseball.checker;
+package baseball.checker.types;
 
 public enum EachNumberResultType {
     STRIKE,
