+++ src/main/java/baseball/players/Players.java
@@ -1,6 +1,5 @@
 package baseball.players;
 
-import java.util.List;
 import java.util.Scanner;
 
 public class Players {
